Service criado mas vou ficar devendo os testes dele

+++ db/migrate/20191112233751_create_movimentacao_estoques.rb
@@ -2,7 +2,7 @@ class CreateMovimentacaoEstoques < ActiveRecord::Migration[5.2]
   def change
     create_table :movimentacao_estoques do |t|
       t.date :data
-      t.integer :quantidade
+      t.float :quantidade
       t.belongs_to :pessoa, foreign_key: true
       t.belongs_to :operacao, foreign_key: true
       t.belongs_to :produto, foreign_key: true
